+++ src/webview/Interfaces/IFeatureUtilities.ts
@@ -1,0 +1,7 @@
+interface IFeatureUtilities {
+    setBounds(bounds: [number, number][]): void;
+    setFeatureName(layer: string, number: number): void;
+    setFeatureDesc(layer: string, number: number): void;
+    panToFeature(layer:string, number: number): void;
+    highlightFeature(layer:string, number: number): void;
+}