+++ src/webview/Interfaces/IIconManager.ts
@@ -1,5 +1,5 @@
 interface IIconManager {
-    getIcon(name: string): string;
+    getIcon(name: string, style:string): string;
     getDivIcon(name: string):  L.DivIcon;
     setIcon(name: string, value:string ): void;
 }