+++ src/webview/Interfaces/IVersionManager.ts
@@ -1,0 +1,4 @@
+interface IVersionManager{
+    version:string
+    convert(content: string): string;
+}