+++ src/webview/MapUtilities.ts
@@ -21,28 +21,28 @@ export module MapUtilities {
         }
 
         /**
-         * Changes the name for the marker at the given index.
+         * Changes the name for the feature at the given index.
          * @param {*} index Index of the edited marker in the markerLayer.
          */
-        setMarkerName(index: number) {
-            let text = (<HTMLInputElement>document.getElementById("markerName" + index))?.value;
-            let marker: any = this.layerManager.getLayer("marker").getLayers()[index];
+        setFeatureName(layer: string, index: number) {
+            let text = (<HTMLInputElement>document.getElementById(layer + "Name" + index))?.value;
+            let feature: any = this.layerManager.getLayer(layer).getLayers()[index];
 
-            marker.feature.properties.name = text;
+            feature.feature.properties.name = text;
 
-            this.setMarkerDesc(index);
+            this.setFeatureDesc(layer, index);
         }
 
         /**
-         * Changes the description and name for the marker at the given index.
+         * Changes the description and name for the feature at the given index.
          * @param {*} index Index of the marker in the markerLayer.
          */
-        setMarkerDesc(index: number) {
-            let text = (<HTMLInputElement>document.getElementById("marker" + index))?.value;
-            let marker: any = this.layerManager.getLayer("marker").getLayers()[index];
+        setFeatureDesc(layer: string, index: number) {
+            let text = (<HTMLInputElement>document.getElementById(layer + index))?.value;
+            let feature: any = this.layerManager.getLayer(layer).getLayers()[index];
 
-            marker.feature.properties.desc = text;
-            marker.getPopup().setContent("<h1>" + marker.feature.properties.name + "</h1>" + "<hr  class=separator>" + text);
+            feature.feature.properties.desc = text;
+            feature.getPopup().setContent("<h1>" + feature.feature.properties.name + "</h1>" + "<hr  class=separator>" + text);
 
             /**
              * Since we modified a layer, the map has to be saved.
@@ -51,13 +51,22 @@ export module MapUtilities {
         }
 
         /**
-         * Pan to the marker with the given index.
+         * Pan to the feature with the given index.
          * @param {*} index Index of the marker in the markerLayer.
          */
-        panToMarker(index: number) {
+        panToFeature(layer: string, index: number) {
 
-            var marker: any = this.layerManager.getLayer("marker").getLayers()[index];
-            this.map.panTo(marker.getLatLng());
+            var feature: any = this.layerManager.getLayer(layer).getLayers()[index];
+
+            /**
+             * The retrieved feature could be any feature type, not all support getLatLng, such as those descending from polyline.
+             * We can use getCenter instead for those.
+             */
+            if (feature.getLatLng) {
+                this.map.panTo(feature.getLatLng());
+            } else if (feature.getCenter) {
+                this.map.panTo(feature.getCenter());
+            }
         }
 
         /**
@@ -65,14 +74,43 @@ export module MapUtilities {
          * Only one marker is kept highlighted at a time, the previous one is reset to its icon.
          * @param index Index of the marker in the marker layer.
          */
-        highlightMarker(index: number){
-            var marker: any = this.layerManager.getLayer("marker").getLayers()[index];
-            this.highlight?.setIcon(this.iconManager.getDivIcon("marker"));
-            this.highlight = marker;
-            this.highlight.setIcon(this.iconManager.getDivIcon("marker-highlight"));
+        highlightFeature(layer: string, index: number) {
+            /**
+             * Current feature which we will highlight.
+             */
+            var feature: any = this.layerManager.getLayer(layer).getLayers()[index];
+
+
+            /**
+             * Reset previous highlight
+             * Currently only markers use icons.
+             */
+            if (this.highlight?.setIcon) {
+
+                this.highlight?.setIcon(this.iconManager.getDivIcon("marker"));
+
+            } else {
+
+                this.highlight?.setStyle({ color: "rgb(99, 161, 255)" });
+            }
+
+            /**
+             * Set new highlight
+             * Not every feature is icon based.
+             */
+            if (feature.setIcon) {
+
+                this.highlight = feature;
+                this.highlight.setIcon(this.iconManager.getDivIcon("marker-highlight"));
+                
+            } else {
+
+                feature.setStyle({ color: "gold" });
+                this.highlight = feature;
+            }
         }
 
-        setBounds (bounds:  [number, number][]){
+        setBounds(bounds: [number, number][]) {
             this.map.fitBounds(bounds);
         }
     }