+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 # Change Log
 
+## 0.2.1
+
+- The Activity Bar entry now has a button for creating new maps.
+- Icons are now uniform.
+- Added popup, text, panning and highlight support for the remaining layers.
+- Editing circles should now correctly save their radius.
+  
+Other:
+
+- HTML files were missing from source, they are copied to the out folder when running the pack command.
+- Moved interfaces to their own folder.
+- Sidebar panels now have general functions for panel creation.
+- Added a class for patching outdated map files in case of breaking changes.
+
 ## 0.2.0
 
-- Initial release
-- Ported to Typescript
-- Added activity bar entry for recently opened maps
+- Initial release.
+- Ported to Typescript.
+- Added activity bar entry for recently opened maps.